Complete cookies API
Work in progress.

See #215.
